Created AnimalCreator class
